Deprecate asdf.asdftypes in favor of asdf.types
